Subject: Improved flexibility of canopy statistics computation
Body: Updated the 'ATL03_ATL08_compute_seg_attributes_dt_segStat' function to make it more robust and flexible.

Previously, the function required specifying photon 'quality' and 'night_flag' parameters, even when they're not needed, which limited flexibility and could lead to unneeded complexity or potential errors in certain use cases.

The function now accepts NULL values for 'quality_ph' and 'night_flag', meaning you can skip filtering on these parameters if they're not relevant for the data you're working with.

Additionally, minor modifications were made to example and comment wording to increase clarity and understanding. We replaced the term "customized function" with the more accurate description "customized list expressions". Moreover, a class has been prepended to the output of the function for consistency in output structure.

These changes aim to improve the usability of the function in a wider variety of situations and increase the clarity of its output and documentation.