ATL03_ATL08_photons_attributes_dt_LAS: Fix UTM zones iteration (seq_along was using number of columns)
Updated function handling Photon attributes LAS conversion. The changes in the code focus on improving the clarity of the function output and efficiency of the iteration process.

To aid readability in the example of usage, a temporary directory variable, `outdir`, has been initialized. The function argument's representation is made more explicit too, replacing `file.path()` with `output = file.path()`. This helps in better understanding the function's API.

The loop now operates over the `epsg` attribute of the `maskZones` which increases efficiency by operating directly on the intended object. The print statement for each local output file has been moved after the file write operation, ensuring files are successfully written before the message is printed, thereby making the debug information more accurate.

Note: This changeset does not modify logic in any significant way and is expected not to affect function outputs or performance.