Added documentation for ATL03_read and ATL08_read methods.
Improved documentation and extended reading capabilities for ICESat-2 ATL03 and ATL08 data in the R package. Added detailed descriptions, parameters, and return values for methods reading from local files and the cloud. This includes handling multiple granules for ATL08, with an explicit stop for multiple granule inputs to ensure the package's current one-granule-at-a-time limitation is clear to users. These enhancements make it easier for users to understand function usage and limitations, providing a better overall user experience when interacting with ICESat-2 data.

No specific issues referenced.